+++ packages/sidebar/src/index.tsx
@@ -5,13 +5,12 @@ import Overlay from '@lendi-ui/overlay';
 import Transition, { State } from '@lendi-ui/transition';
 
 import { Wrapper, SidebarContent, SidebarFooter } from './index.style';
-import { OVERLAY_ZINDEX } from './constants';
 
 export type SidebarDirection = 'left' | 'right';
 
 export interface SidebarCompoundComponent {
-  Content: React.FunctionComponent;
-  Footer: React.FunctionComponent;
+  Content: React.FunctionComponent<SidebarContentProps>;
+  Footer: React.FunctionComponent<SidebarFooterProps>;
 }
 
 export interface SidebarProps extends LUIGlobalProps {
@@ -19,6 +18,7 @@ export interface SidebarProps extends LUIGlobalProps {
   className?: string;
   direction?: SidebarDirection;
   onHide?: () => void;
+  overlayZIndex?: number;
   top?: number;
   isVisible?: boolean;
 }
@@ -29,12 +29,13 @@ const Sidebar: React.FunctionComponent<SidebarProps> & SidebarCompoundComponent 
   top = 0,
   isVisible = false,
   onHide,
+  overlayZIndex = 2, // We just default this value to 2 right now in order to keep consistensy with NavbarBase and update it once LUI z-index management completed.
   className,
   ...otherProps
 }: SidebarProps) => {
   return (
     <>
-      <Overlay isVisible={isVisible} zIndex={OVERLAY_ZINDEX} onHide={onHide} />
+      <Overlay isVisible={isVisible} zIndex={overlayZIndex} onHide={onHide} />
       <Transition isActive={isVisible} timeout={250}>
         {(state: State) => (
           <Wrapper side={direction} transition={state} top={top} className={className} {...otherProps}>
@@ -54,11 +55,11 @@ Sidebar.Content = ({ children }: SidebarContentProps) => <SidebarContent>{childr
 
 export interface SidebarFooterProps {
   children?: React.ReactNode;
-  isTopShadow?: boolean;
+  hasTopShadow?: boolean;
 }
 
-Sidebar.Footer = ({ children, isTopShadow = false }: SidebarFooterProps) => (
-  <SidebarFooter isTopShadow={isTopShadow}>{children}</SidebarFooter>
+Sidebar.Footer = ({ children, hasTopShadow = false }: SidebarFooterProps) => (
+  <SidebarFooter hasTopShadow={hasTopShadow}>{children}</SidebarFooter>
 );
 
 export default Sidebar;