+++ packages/sidebar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog - @lendi-ui/sidebar
 
+## [1.1.0] 2019-11-12 [HUB-hotfix-sidebar](https://creditandfinance.atlassian.net/browse.HUB-hotfix-sidebar)
+### Updated
+- Add one more prop 'overlayZIndex' to Sidebar component.
+- Change Sidebar Footer props 'isTopShadow' to 'hasTopShadow'.
+- Fix Siderbar.footer props type issue.
+
 ## [1.0.0] 2019-11-08 [HUB-675](https://creditandfinance.atlassian.net/browse.HUB-675)
 ### Created
 - Created Sidebar component.